Remove SectionInfoProtocol, close #103
Removes the superfluous `SectionInfoProtocol` to simplify `DataSource`.
There was really no need for this. If clients aren't happy with this, they
can still implement their own `DataSourceProtocol` instead.

- Removes `SectionInfoProtocol`
- Use `Section` in `DataSource` directly

Other changes:
- Move `DataSourceProtocol` to its own file
- Move `FetchedResultsController` to its own file
